@@ -1,13 +1,14 @@
 package com.randomappsinc.studentpicker.dialogs;
 
 import android.content.Context;
-import android.support.annotation.NonNull;
 import android.text.InputFilter;
 import android.text.InputType;
 
 import com.afollestad.materialdialogs.DialogAction;
 import com.afollestad.materialdialogs.MaterialDialog;
 import com.randomappsinc.studentpicker.R;
+
+import androidx.annotation.NonNull;
 
 public class DuplicationDialog {
 
