@@ -1,7 +1,6 @@
 package com.randomappsinc.studentpicker.dialogs;
 
 import android.content.Context;
-import android.support.annotation.NonNull;
 
 import com.afollestad.materialdialogs.DialogAction;
 import com.afollestad.materialdialogs.MaterialDialog;
@@ -9,6 +8,8 @@ import com.randomappsinc.studentpicker.R;
 
 import java.util.ArrayList;
 import java.util.List;
+
+import androidx.annotation.NonNull;
 
 public class MergeNameListsDialog {
 
