@@ -1,7 +1,6 @@
 package com.randomappsinc.studentpicker.dialogs;
 
 import android.content.Context;
-import android.support.annotation.NonNull;
 import android.text.InputFilter;
 import android.text.InputType;
 import android.widget.EditText;
@@ -9,6 +8,8 @@ import android.widget.EditText;
 import com.afollestad.materialdialogs.DialogAction;
 import com.afollestad.materialdialogs.MaterialDialog;
 import com.randomappsinc.studentpicker.R;
+
+import androidx.annotation.NonNull;
 
 public class RenameDialog {
 
